requirements.txt
Django==1.9.7
django-autoslug==1.9.3
django-ckeditor==5.1.1
django-debug-toolbar==1.4
django-extensions==1.6.7
django-filter==0.13.0
et-xmlfile==1.0.1
jdcal==1.2
Markdown==2.6.2
openpyxl==2.3.5
Pillow==3.2.0
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19
